Changed array_infinite_computer.getDisplayIndexEnd to be zero index based instead of handling the conversion for the sake of the array.slice.
Updated bottom spacer height calculation to account for new zero index end.

Guarded against leaking undefined from the binary index search to the component. There was an edge case where if there was an array with one item thats height was less than the available container height the index search would return undefined. This is to be expected because the getDisplayIndexEnd was only looking for the CLOSEST_HIGHER item which did not exist. But leaking this result to the component would cause the bottom spacer to have a height of "NaNpx". So the getDisplayIndexEnd function now protects in this particular case and just returns the last index of the array's items.