Scroll bar issue is fixed
In this update, the unnecessary scrollbar is removed using the auto tag in scroll bar for overflowing y-axis. In current situation if the container height is 600px and the element inside it only using the area under 600 px, the scrollbar is still visible. By this update, the scrollbar will only show when it required for loading elements hight more then the container size.